requirements.txt
-i https://pypi.org/simple
attrs==24.2.0; python_version >= '3.7'
awscli==1.34.29; python_version >= '3.8'
awscli-local==0.22.0
boto3==1.35.29; python_version >= '3.8'
botocore==1.35.29; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cryptography==43.0.1; python_version >= '3.7'
docutils==0.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
flake8==7.1.1; python_full_version >= '3.8.1'
idna==3.10; python_version >= '3.6'
iniconfig==2.0.0; python_version >= '3.7'
jinja2==3.1.4; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
localstack-client==2.6
markupsafe==2.1.5; python_version >= '3.7'
mccabe==0.7.0; python_version >= '3.6'
moto==5.0.16; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.6.1; python_version >= '3.8'
pycodestyle==2.12.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyflakes==3.2.0; python_version >= '3.8'
pytest==8.3.3; python_version >= '3.8'
pytest-mock==3.14.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.35.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
responses==0.25.3; python_version >= '3.8'
rpds-py==0.20.0; python_version >= '3.8'
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
s3transfer==0.10.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==2.2.3; python_version >= '3.10'
werkzeug==3.0.4; python_version >= '3.8'
xmltodict==0.13.0; python_version >= '3.4'
